수정 - login-required 작업/관련 Router 구조 변경
