+++ src/middlewares/login-required.ts
@@ -1,39 +1,12 @@
 import { Request, Response, NextFunction } from 'express';
-import jwt, { JwtPayload } from 'jsonwebtoken';
 
-function loginRequired(req: Request, res: Response, next: NextFunction) {
-  // request 헤더로부터 authorization bearer 토큰을 받음.
-  const userToken = req.headers['authorization']?.split(' ')[1];
-
-  // 이 토큰은 jwt 토큰 문자열이거나, 혹은 "null" 문자열이거나, undefined임.
-  // 토큰이 "null" 일 경우, login_required 가 필요한 서비스 사용을 제한함.
-  if (!userToken || userToken === 'null') {
-    console.log('서비스 사용 요청이 있습니다.하지만, Authorization 토큰: 없음');
-    res.status(403).json({
-      result: 'forbidden-approach',
-      reason: '로그인한 유저만 사용할 수 있는 서비스입니다.',
-    });
-
-    return;
+export const loginRequired = (req: Request, res: Response, next: NextFunction) => {
+  if (!req.user) {
+    const error = new Error('로그인해주세요.');
+    error.name = 'NotAcceptable';
+    res.redirect('/');
+    throw error;
   }
 
-  // 해당 token 이 정상적인 token인지 확인
-  try {
-    const secretKey = process.env.JWT_SECRET_KEY || 'secret-key';
-    const jwtDecoded = jwt.verify(userToken, secretKey) as JwtPayload;
-
-    const userId = jwtDecoded.userId;
-    // req.currentUserId = userId;
-
-    next();
-  } catch (error) {
-    res.status(403).json({
-      result: 'forbidden-approach',
-      reason: '정상적인 토큰이 아닙니다.',
-    });
-
-    return;
-  }
-}
-
-export { loginRequired };
+  next();
+};
