+++ src/routes/api/index.ts
@@ -1,0 +1,2 @@
+export * from './guestbook';
+export * from './user';
