+++ src/server.ts
@@ -3,7 +3,7 @@ import cors from 'cors';
 import passport from 'passport';
 import cookieParser from 'cookie-parser';
 import 'dotenv/config';
-import { guestbookRouter, userRouter, authRouter, chatRouter } from './routes';
+import { apiRouter, authRouter, chatRouter } from './routes';
 import { errorHandler, getUserFromJWT } from './middlewares';
 import { usePassport } from './passport';
 import webSocket from './socket';
@@ -26,8 +26,7 @@ app.use(passport.initialize());
 
 app.use(getUserFromJWT);
 
-app.use('/api', guestbookRouter);
-app.use('/api', userRouter);
+app.use('/api', apiRouter);
 app.use('/auth', authRouter);
 app.use('/chat', chatRouter);
